requirements-test.txt
# Зависимости для тестирования
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0

# Основные зависимости проекта
aiogram>=3.0.0
aiomysql>=0.1.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Дополнительные утилиты для тестирования
faker>=18.0.0
freezegun>=1.2.0

